Split meta data classes into separate file.
